+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - Enh #70: Add `HasY1` trait and corresponding tests for managing SVG `y1` attribute (@terabytesoftw)
 - Enh #71: Add `HasY2` trait and corresponding tests for managing SVG `y2` attribute (@terabytesoftw)
 - Bug #72: Refactor SVG attribute providers to standardize type validation and improve documentation, and remove redundant excessive tests cases (@terabytesoftw)
+- Bug #73: Update attribute retrieval in tests to use `getAttribute()` method for consistency (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024